requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
gunicorn==23.0.0
psycopg[binary]==3.2.3
pytz==2024.2
cloudinary==1.36.0
Pillow==10.1.0